Split out slack-push generation from the scheduling actor
